Tidy up various thread interactions
Introduce the DropWatch utility struct which can be used to watch a
closure and either:
 - execute the closure and then wait for the DropWatch struct to be
   dropped
 - or wait for the struct to be dropped and then execute the closure

It's primarily useful when dealing with Drops that are triggered from an
async context, but could be useful wherever we have a thread that we'd
like to have a managed lifetime.

fixes: #4121
